% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnltt.R
\name{fnltt_singl_acnt}
\alias{fnltt_singl_acnt}
\alias{fnltt_multi_acnt}
\title{Get financial accounts}
\usage{
fnltt_singl_acnt(corp_code, bsns_year, period = c("A", "H", "Q1", "Q3"))

fnltt_multi_acnt(corp_code, bsns_year, period = c("A", "H", "Q1", "Q3"))
}
\arguments{
\item{corp_code}{unique corporate code (single or multi)}

\item{bsns_year}{business year}

\item{period}{period (A: Annual, H: Half year, Q1: 1st Quarter, Q3: 3rd Quarter)}
}
\value{
if not an error, A repsonse \code{data.frame}.
}
\description{
Get information of financial accounts. (2015 ~)
}
\examples{
# get financial accounts (single corporate code)
\dontrun{
fnltt_singl_acnt(corp_code = "00126380", bsns_year = 2023, period = "A")}

# get financial accounts (multiple corporate code)
\dontrun{
fnltt_multi_acnt(corp_code = c("00356370", "00126380"), bsns_year = 2023, period = "A")}

}
