% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnltt.R
\name{fnltt_singl_acnt_all}
\alias{fnltt_singl_acnt_all}
\title{Get financial accounts all}
\usage{
fnltt_singl_acnt_all(
  corp_code,
  bsns_year,
  period = c("A", "H", "Q1", "Q3"),
  fs_div = c("OFS", "CFS")
)
}
\arguments{
\item{corp_code}{unique corporate code}

\item{bsns_year}{business year}

\item{period}{period (A: Annual, H: Half year, Q1: 1st Quarter, Q3: 3rd Quarter)}

\item{fs_div}{type of financial statement (OFS: Original, CFS: Consolidated)}
}
\value{
if not an error, A repsonse \code{data.frame}.
}
\description{
Get information of financial accounts all. (2015 ~)
}
\examples{
# get financial accounts all (single corporate code)
\dontrun{
fnltt_singl_acnt_all(corp_code = "00126380", bsns_year = 2022, period = "A", fs_div = "CFS")}

}
