% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnltt.R
\name{fnltt_singl_indx}
\alias{fnltt_singl_indx}
\alias{fnltt_cmpny_indx}
\title{Get financial indicators}
\usage{
fnltt_singl_indx(
  corp_code,
  bsns_year,
  period = c("A", "H", "Q1", "Q3"),
  index = c("P", "S", "G", "A")
)

fnltt_cmpny_indx(
  corp_code,
  bsns_year,
  period = c("A", "H", "Q1", "Q3"),
  index = c("P", "S", "G", "A")
)
}
\arguments{
\item{corp_code}{unique corporate code (single or multi)}

\item{bsns_year}{business year}

\item{period}{period (A: Annual, H: Half year, Q1: 1st Quarter, Q3: 3rd Quarter)}

\item{index}{financial index (P: Profitability, S: Stability, G: Growth, A: Activity)}
}
\value{
if not an error, A repsonse \code{data.frame}.
}
\description{
Get financial indicators. (2023.3Q ~)
}
\examples{
# get financial indicators (single corporate code)
\dontrun{
fnltt_singl_indx(corp_code = "00126380", bsns_year = 2023, period = "A",
index = "P")}

# get financial indicators (multiple corporate code)
\dontrun{
fnltt_cmpny_indx(corp_code = c("00356370", "00126380"), bsns_year = 2023,
period = "Q3", index = "P")}

}
