% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnltt.R
\name{fnltt_xbrl}
\alias{fnltt_xbrl}
\title{Download original xbrl file of financial statements}
\usage{
fnltt_xbrl(path, rcept_no, period = c("A", "H", "Q1", "Q3"))
}
\arguments{
\item{path}{path to download}

\item{rcept_no}{reception number}

\item{period}{period (A: Annual, H: Half year, Q1: 1st Quarter, Q3: 3rd Quarter)}
}
\value{
A string specifying the file path.
}
\description{
Download original xbrl file of financial statements.
}
\examples{
# download xbrl file
\dontrun{
fnltt_xbrl(rcept_no = "20190401004781", reprt_code = "11011")}

}
