% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corp.R
\name{get_corp_code}
\alias{get_corp_code}
\title{Get unique corporate code data}
\usage{
get_corp_code(path)
}
\arguments{
\item{path}{path to download}
}
\value{
A coporate code \code{data.frame}
}
\description{
Get unique corporate code data
}
