% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corp.R
\name{get_corp_info}
\alias{get_corp_info}
\title{Get corporate basic information}
\usage{
get_corp_info(corp_code)
}
\arguments{
\item{corp_code}{corporate code}
}
\value{
A response \code{data.frame}
}
\description{
Get corporate basic information.
}
\examples{
# get information
\dontrun{
get_corp_info("00126380")}

}
