% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnltt.R
\name{get_xbrl_taxonomy}
\alias{get_xbrl_taxonomy}
\title{Get XBRL taxonomy form}
\usage{
get_xbrl_taxonomy(sj_div)
}
\arguments{
\item{sj_div}{financial statement classification}
}
\value{
A string specifying the file path.
}
\description{
Get XBRL taxonomy form.
}
\examples{
# get xbrl taxonomy form
\dontrun{
fnltt_xbrl(rcept_no = "20190401004781", reprt_code = "11011")}

}
