% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opendart-package.R
\docType{package}
\name{opendart-package}
\alias{opendart}
\alias{opendart-package}
\title{opendart: Open DART API Wrapper (Repository of Korea's Corporate Filings)}
\description{
Anyone can get information of the disclosure reports published on DART
through the open API.

\itemize{
\item{DART disclosure statement}
\item{Key disclosures and financial information}
\item{Large amounts of financial information}
}
}
\author{
\strong{Maintainer}: Seokhoon Joo \email{seokhoonj@gmail.com}

}
\keyword{internal}
