% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key.R
\name{set_api_key}
\alias{set_api_key}
\alias{print_api_key}
\title{Set an open api key}
\usage{
set_api_key(api_key)

print_api_key()
}
\arguments{
\item{api_key}{A string specifying opendart api key}
}
\value{
No return value.
}
\description{
Save opendart api key as an environment variable for the current session. \cr
To set it permanently, please add the following line to your .Renvrion file: \cr

OPENDART_API_KEY="YOUR API KEY" \cr

Please open .Renviron file via following code. \cr

#> usethis::edit_r_environ() \cr
}
\details{
Set an open api key to use opendart.
}
