% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{url_fetch}
\alias{url_fetch}
\title{Fetch URL}
\usage{
url_fetch(api_url, ...)
}
\arguments{
\item{api_url}{url path after base url (\code{ex.} 'fnlttSinglAcnt.json')}

\item{...}{request arguments \describe{
\itemize{
\item{\code{corp_code}}
\item{\code{bsns_year}}
\item{\code{reprt_code}}
\item{\code{idx_cl_code}}
\item{\code{bgn_de}}
\item{\code{end_de}}
\item{\code{fs_div}}
\item{\code{sj_div}}
}}}
}
\value{
if not an error, A repsonse \code{data.frame}.
}
\description{
Fetch URL
}
\examples{
# example code
\dontrun{
url_fetch(api_url = "fnlttSinglIndx.json", corp_code = "00126380",
bsns_year = 2023, reprt_code = "11014", idx_cl_code = "M210000")}

}
